demo-admin-backend/requirements.txt
amqp==2.5.1
aniso8601==7.0.0
APScheduler==3.6.3
atomicwrites==1.3.0
attrs==19.1.0
aws-requests-auth==0.4.2
billiard==3.6.0.0
blist==1.3.6
boto3==1.10.18
botocore==1.13.18
celery==4.3.0
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
configparser==4.0.2
confluent-kafka==1.5.0
croniter==0.3.30
defusedxml==0.6.0
docopt==0.6.2
docutils==0.15.2
elastalert==0.2.4
elasticsearch==7.0.0
envparse==0.2.0
exotel==0.1.5
Flask==1.0.2
Flask-Cors==3.0.7
flask-mongoengine==0.9.5
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
future==0.18.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
jira==2.0.0
jmespath==0.9.4
jsonschema==3.2.0
kombu==4.6.4
MarkupSafe==1.1.1
marshmallow==2.15.6
mock==3.0.5
mongoengine==0.18.2
more-itertools==7.2.0
oauthlib==3.1.0
pbr==5.4.3
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.1
PySocks==1.7.1
PyStaticConfiguration==0.10.4
pytest==3.8.0
pytest-mock==1.10.0
python-dateutil==2.6.1
python-dotenv==0.10.3
python-magic==0.4.15
python-slugify==4.0.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.12.0
stomp.py==4.1.22
text-unidecode==1.3
texttable==1.6.2
thehive4py==1.6.0
twilio==6.0.0
tzlocal==2.0.0
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.5
WTForms==2.2.1
zipp==0.6.0
